перенёс создание комментариев в другой модуль
