+++ js/windowRendering.js
@@ -1,6 +1,10 @@
 import {
   photosData
 } from './createArrayPhotoData.js';
+import {
+  createComments,
+  socialCommentsFragment
+} from './renderComments.js';
 
 //Нужно чтобы окно открывалось при нажатии на миниматюру данные берём из photosData;
 const body = document.querySelector('body');
@@ -11,26 +15,27 @@ const bigPicturelikes = document.querySelector('.likes-count');
 const bigPictureComments = document.querySelector('.social__comment-shown-count');
 const bigPictureCommentsAll = document.querySelector('.social__comment-total-count');
 const commentsNode = document.querySelector('.social__comments');
-const commentsNodeTemplate = commentsNode.querySelector('.social__comment');
+
 const commentsNodeCounter = document.querySelector('.social__comment-count');
 const commentsNodeLoader = document.querySelector('.comments-loader');
 
-//Создаём фрагмент для комментария
-const socialCommentsFragment = document.createDocumentFragment();
 
-//Функция создания комментария
-const createComments = (photoData) => {
-  photoData.comments.forEach((comment) => {
-    const socialCommentNode = commentsNodeTemplate.cloneNode(true);
+//Функция закрытия фото.
+const closePhoto = () => {
+  bigPictureContainer.classList.add('hidden');
+  body.classList.remove('modal-open');
 
-    socialCommentNode.querySelector('.social__picture').src = comment.avatar;
-    socialCommentNode.querySelector('.social__picture').alt = comment.name;
-    socialCommentNode.querySelector('.social__text').textContent = comment.message;
-    socialCommentsFragment.appendChild(socialCommentNode);
-  });
+  document.removeEventListener('keydown', stringEscOnDocument);
 };
+//функция для закрытия по ESC
+function stringEscOnDocument(evt) {
+  if (evt.key === 'Escape') {
+    evt.preventDefault();
+    closePhoto();
+  }
+}
 
-//функциz открытия большого изображения
+//функция открытия большого изображения
 const openPhoto = (photoData) => {
   bigPictureContainer.classList.remove('hidden');
   bigPictureImage.src = photoData.url;
@@ -38,26 +43,15 @@ const openPhoto = (photoData) => {
   commentsNode.innerHTML = '';
   createComments(photoData);
   commentsNode.appendChild(socialCommentsFragment);
-
   bigPictureComments.textContent = photoData.comments.length;
   bigPictureCommentsAll.textContent = photoData.comments.length;
   commentsNodeCounter.classList.add('hidden');
   commentsNodeLoader.classList.add('hidden');
+  //Закртытие фото по кнопке.
+  document.addEventListener('keydown', stringEscOnDocument);
   body.classList.add('modal-open');
 };
-//Функция закрытия фото.
-const closePhoto = () => {
-  bigPictureContainer.classList.add('hidden');
-  body.classList.remove('modal-open');
-};
-//Закртытие фото по кнопке.
-document.addEventListener('keydown', (evt) => {
-  if (evt.key === 'Escape') {
-    evt.preventDefault();
-    bigPictureContainer.classList.add('hidden');
-    body.classList.remove('modal-open');
-  }
-});
+
 //Закрытие фото при клике.
 bigPictureClose.addEventListener('click', closePhoto);
